backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.1.0
pyodbc==4.0.32
tensorflow==2.7.0
Pillow==8.4.0
